Maybe not necessary, but I think we should limit the visibility of operations interacting with the Span type now: - Refactor ISpan-compatible extensions into a separate (public) ISpanExtensions class - Reduce visibility of SpanExtensions.SetTraceSamplingPriority to internal
