+++ tracer/src/Datadog.Trace/ExtensionMethods/ISpanExtensions.cs
@@ -1,0 +1,83 @@
+﻿// <copyright file="ISpanExtensions.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Common;
+using Datadog.Trace.Headers;
+using Datadog.Trace.Tagging;
+using Datadog.Trace.Util;
+using Datadog.Trace.Vendors.Serilog;
+
+namespace Datadog.Trace.ExtensionMethods
+{
+    /// <summary>
+    /// Extension methods for the <see cref="ISpan"/> class.
+    /// </summary>
+    // ReSharper disable once InconsistentNaming
+    public static class ISpanExtensions
+    {
+        /// <summary>
+        /// Adds standard tags to a span with values taken from the specified <see cref="DbCommand"/>.
+        /// </summary>
+        /// <param name="span">The span to add the tags to.</param>
+        /// <param name="command">The db command to get tags values from.</param>
+        public static void AddTagsFromDbCommand(this ISpan span, IDbCommand command)
+        {
+            span.ResourceName = command.CommandText;
+            span.Type = SpanTypes.Sql;
+
+            var tags = DbCommandCache.GetTagsFromDbCommand(command);
+
+            foreach (var pair in tags)
+            {
+                span.SetTag(pair.Key, pair.Value);
+            }
+        }
+
+        internal static void DecorateWebServerSpan(
+            this ISpan span,
+            string resourceName,
+            string method,
+            string host,
+            string httpUrl,
+            WebTags tags,
+            IEnumerable<KeyValuePair<string, string>> tagsFromHeaders)
+        {
+            span.Type = SpanTypes.Web;
+            span.ResourceName = resourceName?.Trim();
+
+            tags.HttpMethod = method;
+            tags.HttpRequestHeadersHost = host;
+            tags.HttpUrl = httpUrl;
+
+            foreach (KeyValuePair<string, string> kvp in tagsFromHeaders)
+            {
+                span.SetTag(kvp.Key, kvp.Value);
+            }
+        }
+
+        internal static void SetHeaderTags<T>(this ISpan span, T headers, IReadOnlyDictionary<string, string> headerTags, string defaultTagPrefix)
+            where T : IHeadersCollection
+        {
+            if (headerTags is not null && !headerTags.IsEmpty())
+            {
+                try
+                {
+                    var tagsFromHeaders = SpanContextPropagator.Instance.ExtractHeaderTags(headers, headerTags, defaultTagPrefix);
+                    foreach (KeyValuePair<string, string> kvp in tagsFromHeaders)
+                    {
+                        span.SetTag(kvp.Key, kvp.Value);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Log.Error(ex, "Error extracting propagated HTTP headers.");
+                }
+            }
+        }
+    }
+}
